feat: build nvidia images in main repo
This adds the nvidia build scripts to the repo and reconfigures
Containerfile to build both a main and nvidia image per provided
build-args. Workflow is also adapted with a matrix to support this use
case. "main" build and post-install scripts renamed to make things more
consistent and easier to follow for future dev. COPY used instead of ADD
in Containerfile for best practice compliance.