feat: allow rescue/emergency boot with grub cmdline args
This uses a systemd-generator to dyamically write a drop-in config for
the rescue and emergency services only when they are requested via the
kernel cmdline, which requires console/grub access. This allows use of
these modes with the default Fedora state of a password locked root
user, but does not auto-allow root access in the case of a failed
fsck-check, which can also drop into the emergency shell.

Relates: #470