chore(ci): improve retry errors for github installs
A helpful issue was filed with PR which will help address some spurious
issues with the github actions workflows. That inspired me to improve
the github-release-install.sh script such that it will more properly
fail(retry) when http errors occur.

In addition, this includes an improvement to the script which allows
installing specific tags, not just the latest release.

Relates: #502