chore: adds the alternatives workaround
This creates /var/lib/alternatives so that any RPM which has a
post-install script referencing 'alternatives' will not fail due to its
absense.

In addition, this goes the extra mile by copying the contents of
/var/lib/alternatives into the resulting image, which preserves the
altenratives database for any installed packages.

This second part could be controversial as there is an expectation that
there's minimal (or no) contents left in /var after running
'ostree container commit', but it does work and is useful.