chore: revert change of gstreamer-plugins
In PR #636 I was over-eager in changing out "-free" packages which had
alternatives in negativo17's fedora-multimedia repo.

Specifically, the gstreamer-plugins provided have a large set of
dependencies which increase the size of the "*-main" images by
approximately 1GiB. Reverting that change as if a user or downstream
wishes it, they can make the change (eg, Bazzite does this I think), but
others should not be burdened with this size change, expecially not as
part of a PR which was just supposed to swap out the source of some
packages.