fix: mitigate OpenCL packaging error
It seems a new package was created in Fedora release repos which has
started to be installed in all builds of Fedora 41/42, but it was not
meant to be installed. The new OpenCL-ICD-Loader conflicts with ocl-icd
which is required by mesa-OpenCL, nvidia drivers, etc.

This mitigation simply ensures that ocl-icd is installed, not the new,
undesired package.

Upstream bug report: https://bugzilla.redhat.com/show_bug.cgi?id=2332429