chore: cleanup packages.sh
packages.sh and packages.json are designed around the behavior of
rpm-ostree. Namely, it can install multiple packages while removing
multiple packages in a single transaction. `dnf5` does not support this.
`dnf5 distro-sync` can bring multiple packages into sync at once, but
cannot be used to install packages. `dnf5 swap` only works on one
package at a time.

This cleanup removes the behaviour of doing a single transaction of
installing and removing. Instead, it parses the the json file to install
packages first. Then, it parses the json file and confirms the packages
are on image before removing them.

Already merged is a basic package checker for some of the major packages
for each image (F42 images only. Lazurite and Vauxite are on their own
right now.)

This is a short term change right now that slightly changes
functionality. We already are installing items in install.sh seperate
from this json being the ultimate source of truth.

Signed-off-by: m2 <69128853+m2Giles@users.noreply.github.com>