fix: versionlock QT6
This prevents QT package skew, which may crash SDDM/KWin
making it impossible to login graphically.

This could happen in scenarios where the base image
is still on an old QT version, while the Fedora repositories
have newer versions of the QT packages available.

This happened a few days ago when fcitx* was updated
and the installation of those packages requested the newest version
of QT packages, which caused a mismatch between QT and fcitx packages.

More info:
https://gist.github.com/renner0e/f10300de6abe1c58684262378c0d282a