fix: use libheif from negativo for dolphin thumbnails
fixes: https://github.com/ublue-os/bazzite/issues/2985

This package from negativo is needed for having thumbnails in dolphin for
heic and heif images as it ships a functioning version of libheif,
unlike Fedora.

My guess is that it broke on kinoite only (and not on Silverblue) because of
differences in dependency resolution of rpm-ostree and dnf.
We definitively used to ship the negativo version of the package on
Kinoite in the past.

Nautilus file previews continued to work just fine because we didn't
ran into this weird packaging issue on Silverblue.