chore: Add myself to emeritus.md
Universal Blue does not align with my values and what I want out of it anymore. I'd like to emeritus myself and request to be removed from the organization whenever necessary. I'll be working on my own projects and things that align better with my vision of what should be done.